Move entitlement, license key, and SSO logic to enterprise package
