+++ packages/back-end/src/services/auth/OpenIdAuthConnection.ts
@@ -8,6 +8,7 @@ import {
 } from "openid-client";
 import jwtExpress from "express-jwt";
 import jwks from "jwks-rsa";
+import { SSO_CONFIG } from "enterprise";
 import { AuthRequest } from "../../types/AuthRequest";
 import { MemoryCache } from "../cache";
 import {
@@ -15,7 +16,7 @@ import {
   UnauthenticatedResponse,
 } from "../../../types/sso-connection";
 import { AuthChecksCookie, SSOConnectionIdCookie } from "../../util/cookie";
-import { APP_ORIGIN, IS_CLOUD, SSO_CONFIG } from "../../util/secrets";
+import { APP_ORIGIN, IS_CLOUD } from "../../util/secrets";
 import { getSSOConnectionById } from "../../models/SSOConnectionModel";
 import {
   getAuditableUserPropertiesFromRequest,
