+++ packages/enterprise/test/enterprise.test.ts
@@ -1,7 +1,48 @@
-import { returnTrue } from "../src";
+import { getAccountPlan } from "../src";
 
-describe("enterprise", () => {
-  it("runs tests", () => {
-    expect(returnTrue()).toEqual(true);
-  });
+describe("determines account plan from org", () => {
+  expect(getAccountPlan({})).toBe(false);
+  expect(
+    getAccountPlan({
+      enterprise: true,
+    })
+  ).toBe(true);
+  expect(
+    getAccountPlan({
+      restrictAuthSubPrefix: "something",
+    })
+  ).toBe(true);
+  expect(
+    getAccountPlan({
+      restrictLoginMethod: "something",
+    })
+  ).toBe(true);
+  expect(
+    getAccountPlan({
+      subscription: {
+        status: "canceled",
+      },
+    })
+  ).toBe(false);
+  expect(
+    getAccountPlan({
+      subscription: {
+        status: "active",
+      },
+    })
+  ).toBe(true);
+  expect(
+    getAccountPlan({
+      subscription: {
+        status: "trialing",
+      },
+    })
+  ).toBe(true);
+  expect(
+    getAccountPlan({
+      subscription: {
+        status: "past_due",
+      },
+    })
+  ).toBe(true);
 });
