+++ packages/front-end/components/Layout/SidebarLink.tsx
@@ -3,8 +3,9 @@ import Link from "next/link";
 import { IconType } from "react-icons/lib";
 import { useRouter } from "next/router";
 import clsx from "clsx";
+import { AccountPlan } from "enterprise";
 import { FiChevronRight } from "react-icons/fi";
-import { AccountPlan, Permission } from "back-end/types/organization";
+import { Permission } from "back-end/types/organization";
 import { useGrowthBook } from "@growthbook/growthbook-react";
 import { AppFeatures } from "@/types/app-features";
 import { isCloud } from "../../services/env";
